final commit for tonight

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("Size of a char: %i byte(s)\n", (unsigned long)sizeof(char));
-printf("Size of an int: %i byte(s)\n", (unsigned long)sizeof(int));
-printf("size of a long int: %i bytes(s)\n", (unsigned long)sizeof(long int));
-printf("Size of long int: %i byte(s)\n", (unsigned long)sizeof(long long int));
-printf("Size of a float: %i bytes(s)", (unsigned long)sizeof(float));
+printf("Size of a char: %i byte(s)\n", sizeof(char));
+printf("Size of an int: %i byte(s)\n", sizeof(int));
+printf("size of a long int: %i bytes(s)\n", sizeof(long int));
+printf("Size of long int: %i byte(s)\n", sizeof(long long int));
+printf("Size of a float: %i bytes(s)", sizeof(float));
 return (0);
 }
